Leerkes {WRS2}	R Documentation

Maternal Self-Efficacy

Description

TIn this dataset (n = 92) the relationship between how girls were raised by there own mother and their later feelings of maternal self-efficacy, i.e. our belief in our ability to succeed in specific situations. The other variable is self-esteem (can act as mediator). All variables are scored on a continuous scale from 1 to 4.

Usage

Leerkes

Format

A data frame with 3 variables and 92 observations:

MatCare

maternal care

Efficacy

maternal self-efficacy

Esteem

self-esteen

References

Leerkes, E.M. & Crockenberg, S.C. (2002). The development of maternal self-efficacy and its impact on maternal behavior. Infancy, 3, 227-247.

Howell, D.C. (2012). Statistical Methods for Psychology (8th edition). Wadsworth, Belmont, CA.

Examples

summary(Leerkes)

[Package WRS2 version 1.1-7 Index]