WineTasting {WRS2}	R Documentation

Wine Tasting

Description

In this hypothetical dataset we have three types of wine (A, B and C). We asked 22 friends to taste each of the three wines (in a blind fold fashion), and then to give a grade of 1 to 7. We asked them to rate the wines 5 times each, and then averaged their results to give a number for a persons preference for each wine.

Usage

WineTasting

Format

A data frame with 3 variables and 66 observations:

Taste

Taste Rating

Wine

Wine (A, B, C)

Taster

Taster (index)

Examples

WineTasting
summary(WineTasting)

[Package WRS2 version 1.1-7 Index]