bush {WRS2}	R Documentation

Bushtucker Foods

Description

In the TV show "I'm a celebrity, get me out of here" the celebrities had to eat things like stick insects, fish eyes, etc. This dataset records the time taken to retch when eating these things.

Usage

bush

Format

A data frame with 5 variables and 8 observations:

participant

participant ID

stick_insect

time taken to retch when eating a stick insect

kangaroo_testicle

time taken to retch when eating a kangaroo testicle

fish_eye

time taken to retch when eating a fish eye

witchetty_grub

time taken to retch when eating a witchetty grub

Details

Dataset from Field et al. book (p. 557).

References

Field, A., Miles, J., & Field, Z. (2012). Discovering Statistics Using R. Sage.

Examples

bush
summary(bush)

[Package WRS2 version 1.1-7 Index]