electric {WRS2}	R Documentation

The Electric Company

Description

These data are based on an educational TV show for children called "The Electric Company". In each of four grades, the classes were randomized into treated (TV show) and control groups (no TV show). At the beginning and at the end of the school year, students in all the classes were given a reading test. The average test scores per class were recorded.

Usage

electric

Format

A data frame with 5 variables and 192 observations:

City

Fresno and Youngstown

Grade

grade

Pretest

reading scores at the beginning of the semester

Posttest

reading scores at the end of the semester

Group

treatment vs. control

References

Gelman, A., & Hill, J. (2007). Data Analysis Using Regression and Multilevel/Hierarchical Models. Cambridge University Press: New York, NY.

Examples

summary(electric)

[Package WRS2 version 1.1-7 Index]