essays {WRS2}	R Documentation

Academic Writing Data

Description

This study looked at the effects of two forms of written corrective feedback on lexico-grammatical accuracy in the academic writing of English as a foreign language university students. It had a 3x4 within-by-between design with three groups (two treatment and one control) measured over four occasions (pretest/treatment, treatment, posttest, delayed posttest).

Usage

bush

Format

A data frame with 4 variables and 120 observations:

id

participant ID

group

control, direct, indirect

essay

four measurement occasions

errorRatio

error ratio

References

McGrath, D. (2016). The Effects of Comprehensive Direct and Indirect Written Corrective Feedback on Accuracy in English as a Foreign Language Students' Writing (Unpublished master's thesis). Macquarie University, Sydney, Australia.

Examples

head(essays)
summary(essays)

[Package WRS2 version 1.1-7 Index]