goggles {WRS2}	R Documentation

Beer Goggles Effect

Description

This dataset is about the effects of alcohol on mate selection in night-clubs. The hypothesis is that after alcohol had been consumed, subjective perceptions of physical attractiveness would become more inaccurate (beer-goggles effect). There are 48 participants: 24 males, 24 females. The reseacher took 3 groups of 8 participants to a night club. One group got no alcohol, one group 2 pints, and one group 4 pints. At the end of the evening the researcher took a photograph of the person the participant was chatting up. The attractiveness of the person on the photo was then evaluated by independent judges.

Usage

goggles

Format

A data frame with 3 variables and 48 observations:

gender

24 male, 24 female students

alcohol

amount of alcohol consumed

attractiveness

attractiveness rating (0-100)

Details

Dataset from Field et al. book (p. 501).

References

Field, A., Miles, J, & Field, Z. (2012). Discovering Statistics Using R. Sage.

Examples

goggles
summary(goggles)

[Package WRS2 version 1.1-7 Index]