hangover {WRS2}	R Documentation

Hangover Symptoms

Description

In a study on the effect of consuming alcohol, hangover symptoms were measured for two independent groups, with each subject consuming alcohol and being measured on three different occasions. One group consisted of sons of alcoholics and the other was a control group.

Usage

hangover

Format

A data frame with 4 variables and 120 observations:

symptoms

number of hangover symptoms

group

son of alcoholic vs. control

time

measurement occasion

id

subject ID

References

Wilcox, R. (2012). Introduction to Robust Estimation and Hypothesis Testing (3rd ed.). Elsevier.

Examples

summary(hangover)

[Package WRS2 version 1.1-7 Index]