med1way {WRS2}	R Documentation

Heteroscedastic one-way ANOVA for medians.

Description

This function computes a one-way ANOVA for the medians. Homoscedasticity assumption not required. There shouldn't be too many ties.

Usage

med1way(formula, data, iter = 1000, ...)

Arguments

formula
an object of class formula.

data
an optional data frame for the input data.

iter
number of iterations to determine critical value.

...
currently ignored.

Details

Evaluating the test statistic using the df proposed in the literature can result in the actual level being less than the nominal level, (i.e., around 0.02-0.025 when testing at the 0.05 level and the sample size is small). A better strategy is to simulate the critical value and computing the p-value accordingly, as implemented in this function.

Value

Returns an object of class med1way containing:

test
F-value of the test statistic

crit.val
critical value

p.value
p-value

call
function call

References

Wilcox, R. (2012). Introduction to Robust Estimation and Hypothesis Testing (3rd ed.). Elsevier.

See Also

t1way,t1waybt, Qanova

Examples

set.seed(123)
med1way(libido ~ dose, data = viagra, iter = 3000)

[Package WRS2 version 1.1-7 Index]