med2way {WRS2}	R Documentation

A two-way ANOVA for medians.

Description

This function computes a two-way ANOVA medians with interactions effects.

Usage

med2way(formula, data, ...)

Arguments

formula
an object of class formula.

data
an optional data frame for the input data.

...
currently ignored.

Details

There should not be too many ties in the data. The test statistics for the main effects in med2way are F-distributed, the (heteroscedastic) test for the interaction is chi-square distributed. Post hoc tests can be performed using mcp2a.

Value

Returns an object of class t2way containing:

Qa
first main effect

A.p.value
p-value first main effect

Qb
second main effect

B.p.value
p-value second main effect

Qab
interaction effect

AB.p.value
p-value interaction effect

call
function call

varnames
variable names

dim
design dimensions

References

Wilcox, R. (2012). Introduction to Robust Estimation and Hypothesis Testing (3rd ed.). Elsevier.

See Also

t2way, med1way

Examples

med2way(attractiveness ~ gender*alcohol, data = goggles)
mcp2a(attractiveness ~ gender*alcohol, data = goggles, est = "median")

[Package WRS2 version 1.1-7 Index]