movie {WRS2}	R Documentation

Movies and Aggressive Affect

Description

Participants are randomly assigned to one of two groups. The first group watches a violent film, and the other watches a nonviolent film. Afterwards, the aggressive affect is measured, and it is desired to compare three groups, taking gender and degress into account as well.

Usage

movie

Format

A data frame with 4 variables and 68 observations:

degree

no degree vs. degree

gender

36 males, 32 females

type

violend vs. nonviolent

aggressive

aggressive affect

Details

Artificial dataset from Wilcox book (p. 316).

References

Wilcox, R. (2012). Introduction to Robust Estimation and Hypothesis Testing (3rd ed.). Elsevier.

Examples

movie
summary(movie)

[Package WRS2 version 1.1-7 Index]