picture {WRS2}	R Documentation

Profile Pictures

Description

This dataset examines how the profile pictures on social network platforms affect the number of friend requests when females are in a relationship. The relationship status is a between-subject variable (part of the participants did set their status to relationship). For the first 3 weeks the subjects had a picture of their own in their profiles. For the following 3 weeks they posted a picture with a man.

Usage

picture

Format

A data frame with 4 variables and 40 observations:

case

subject id

relationship_status

Relationship status on social network platform

couple

amount of friend requests when profile picture as couple

alone

amount of friend requests when profile picture as single

Details

Dataset from Field et al. book (p. 644).

References

Field, A., Miles, J., & Field, Z. (2012). Discovering Statistics Using R. Sage.

Examples

picture
summary(picture)

[Package WRS2 version 1.1-7 Index]