Qanova {WRS2}	R Documentation

Quantile ANOVA

Description

One-way ANOVA based on quantiles. Only known method to work well when tied values are likely to occur.

Usage

Qanova(formula, data, q = 0.5, nboot = 600, ...)

Arguments

formula
an object of class formula.

data
an optional data frame for the input data.

q
quantile (or vector of quantiles) to be used.

nboot
number of bootstrap samples

...
currently ignored.

Details

Test global hypothesis that J independent groups have equal quantiles (default: median) using the Harrell-Davis estimator. Performs well when there are tied values.

Value

Qanova an object of class "qanova" containing:

psihat
value of the test statistics

contrasts
contrasts

p.value
p-value

call
function call

References

Wilcox, R. (2012). Introduction to Robust Estimation and Hypothesis Testing (3rd ed.). Elsevier.

See Also

qcomhd, t1way, discANOVA

Examples

## median comparison
set.seed(123)
fitmed <- Qanova(libido ~ dose, viagra, nboot = 200)
fitmed

## 1st, 3rd quartile comparison
set.seed(123)
fitquart <- Qanova(libido ~ dose, viagra, q = c(0.25, 0.75), nboot = 200)
fitquart

[Package WRS2 version 1.1-7 Index]