rmanova {WRS2}	R Documentation

A heteroscedastic one-way repeated measures ANOVA for trimmed means.

Description

The rmanova function computes a one-way repeated measures ANOVA for the trimmed means. Homoscedasticity assumption not required. Corresponding post hoc tests can be performed using rmmcp.

Usage

rmanova(y, groups, blocks, tr = 0.2, ...)
rmmcp(y, groups, blocks, tr = 0.2, alpha = 0.05, ...)

Arguments

y
a numeric vector of data values (response).

groups
a vector giving the group of the corresponding elements of y.

blocks
a vector giving the block of the corresponding elements of y.

tr
trim level for the mean.

alpha
alpha level for post hoc comparisons.

...
currently ignored.

Value

rmanova an object of class "t1way" containing:

test
value of the test statistic

df1
degrees of freedom

df2
degrees of freedom

p.value
p-value

call
function call

rmmcp returns an object of class "mcp1" containing:

comp
inference for all pairwise comparisons

fnames
names of the factor levels

References

Wilcox, R. (2012). Introduction to Robust Estimation and Hypothesis Testing (3rd ed.). Elsevier.

See Also

rmanovab, med1way, t1way

Examples

head(WineTasting)
rmanova(WineTasting$Taste, WineTasting$Wine, WineTasting$Taster)

## post hoc
rmmcp(WineTasting$Taste, WineTasting$Wine, WineTasting$Taster)

head(bush)
require(reshape)
bushLong <- melt(bush, id.var = "participant", variable_name = "food")
rmanova(bushLong$value, bushLong$food, bushLong$participant)

## post hoc
rmmcp(bushLong$value, bushLong$food, bushLong$participant)

[Package WRS2 version 1.1-7 Index]