rmanovab {WRS2}	R Documentation

A heteroscedastic one-way repeated measures bootstrap ANOVA for trimmed means.

Description

The rmanovab function computes a bootstrap version of the one-way repeated measures ANOVA for the trimmed means. Homoscedasticity assumption not required. Corresponding post hoc tests can be performed using pairdepb.

Usage

rmanovab(y, groups, blocks, tr = 0.2, nboot = 599, ...)
pairdepb(y, groups, blocks, tr = 0.2, nboot = 599, ...)

Arguments

y
a numeric vector of data values (response).

groups
a vector giving the group of the corresponding elements of y.

blocks
a vector giving the block of the corresponding elements of y.

tr
trim level for the mean.

nboot
number of bootstrap samples.

...
currently ignored.

Value

rmanovab an object of class "rmanovab" containing:

test
value of the test statistic

crit
critical value

call
function call

pairdepb returns an object of class "mcp2" containing:

comp
inference for all pairwise comparisons

fnames
names of the factor levels

References

Wilcox, R. (2012). Introduction to Robust Estimation and Hypothesis Testing (3rd ed.). Elsevier.

See Also

rmanova

Examples

set.seed(123)
rmanovab(WineTasting$Taste, WineTasting$Wine, WineTasting$Taster, nboot = 300)

## post hoc
set.seed(123)
pairdepb(WineTasting$Taste, WineTasting$Wine, WineTasting$Taster, nboot = 300)

[Package WRS2 version 1.1-7 Index]