spider {WRS2}	R Documentation

Arachnophobes

Description

24 arachnophobes were used in all. 12 were asked to play with a big hairy tarantula spider with big fangs and an evil look. Their subsequent anxiety was measured. The remaining 12 were shown only picture of the same hairy tarantula. Again, the anxiety was measured.

Usage

spider

Format

A data frame with 2 variables and 24 observations:

Group

picture vs. real spider

Anxiety

anxiety measure

Details

Dataset from Field et al. book (p. 362).

References

Field, A., Miles, J., & Field, Z. (2012). Discovering Statistics Using R. Sage.

Examples

spider

[Package WRS2 version 1.1-7 Index]