t1way {WRS2}	R Documentation

A heteroscedastic one-way ANOVA for trimmed means.

Description

The t1way function computes a one-way ANOVA on trimmed means. Homoscedasticity assumption not required. It uses a generalization of Welch's method. Corresponding post hoc tests can be performed using lincon.

Usage

t1way(formula, data, tr = 0.2, alpha = 0.05, nboot = 100, ...)
lincon(formula, data, tr = 0.2, alpha = 0.05, method = "hochberg", ...)

Arguments

formula
an object of class formula.

data
an optional data frame for the input data.

tr
trim level for the mean.

alpha
alpha level for CI computation.

nboot
number of bootstrap samples for effect size CI computation.

method
method to correct the p-value (see p.adjust)

...
currently ignored.

Details

In the post hoc computations, confidence intervals and p-values are adjusted to control FWE. The default for the p-values is to use Hochberg's 1988 sharper Bonferroni procedure.

Value

t1way returns an object of class "t1way" containing:

test
value of the test statistic (F-statistic)

df1
degrees of freedom

df2
degrees of freedom

p.value
p-value

effsize
explanatory measure of effect size

effsize_ci
boostrap effect size CI

call
function call

lincon returns an object of class "mcp1" containing:

comp
inference for all pairwise comparisons

fnames
names of the factor levels

linconv2 returns an object of class "linconv2" containing:

References

Wilcox, R. (2012). Introduction to Robust Estimation and Hypothesis Testing (3rd ed.). Elsevier.

See Also

med1way,t1waybt

Examples

set.seed(123)
t1way(libido ~ dose, data = viagra)

## post hoc tests
lincon(libido ~ dose, data = viagra)

[Package WRS2 version 1.1-7 Index]