viagra {WRS2}	R Documentation

Effects of Viagra

Description

Participants were assigned randomly to three viagra dosages (placebo, low dosage, high dosage). The dependent variable was an objective measure of libido.

Usage

viagra

Format

A data frame with 2 variables and 15 observations:

dose

viagra dosage

libido

objective measure of libido

Details

Artificial dataset from Field et al. book (p. 401).

References

Field, A., Miles, J., & Field, Z. (2012). Discovering Statistics Using R. Sage.

Examples

viagra

[Package WRS2 version 1.1-7 Index]