wmcpAKP {WRS2}	R Documentation

Effect size for dependent samples ANOVA

Description

Compute an AKP-type effect size for dependent sample ANOVA

Usage

wmcpAKP(x, tr = 0.2, nboot = 200, ...)

Arguments

x
data frame in wide format (no missing values allowed)

tr
trim level for the means.

nboot
number of bootstrap samples.

...
currently ignored.

Details

The computation is based on a modification of the Algina-Keselman-Penfield effect size for J dependent samples.

References

Algina, J., Keselman, H.J., & Penfield, R.D. (2005). An alternative to Cohen's standardized mean difference effect size: A robust parameter and confidence interval in the two independent groups case. Psychological Methods, 10, 317-328.

See Also

rmanova, yuend

Examples

## Not run: 
require(reshape)
WineLong <- cast(WineTasting, Taster ~ Wine, value = "Taste")[,-1]
set.seed(123)
effsize <- wmcpAKP(WineLong, nboot = 20)
effsize

## End(Not run)

[Package WRS2 version 1.1-7 Index]